Reviews Manager
Refactor the reviews manager.
